update type Interger

+++ src/main/java/vn/framgia/dao/IUserDAO.java
@@ -6,6 +6,6 @@ import vn.framgia.model.User;
  * Created by FRAMGIA\duong.van.tien on 06/03/2017.
  *
  */
-public interface IUserDAO extends IGenericDAO<User, Long>{
+public interface IUserDAO extends IGenericDAO<User, Integer>{
 	public User getUserByAcount(String username);
 }
